requirements.txt
aiomisc==12.0.0
setproctitle==1.2.2
aiohttp==3.7.3
ConfigArgParse==1.3
aio_pika==6.7.1
PyJWT==2.0.1
gunicorn==20.0.4
